Add Invite and Contact list modules
